move get_sys_interfaces to ip_control
